+++ feedme/react/urls.py
@@ -1,8 +1,8 @@
 from django.conf.urls import patterns, include, url
 
+from feedme.react.views import react_index
 
-# react
-from feedme.react_views import react_index
+
 urlpatterns = (
     url(r'^$', react_index, name='react_index'),
 )
